net/wr-mch: support for RX, TX timestamps
The endpoint tags incoming packets and appends the timestamp
to the corresponding field in the WR header.
TX packets are trickier: a tag is put into a field and then
at the endpoint this tag is associated to a timetag. This
pair is stored in a FIFO and we retrieve it from there
after receiving the TX done interrupt--since the endpoint
puts the tag into the FIFO when the transfer through it
initiates, we know that when the transfer finishes the
pair timestamp-tag is going to be there to be read. Anyway
if it's not there yet, we silently bail out (since we're
doing this in the interrupt handler).

Signed-off-by: Emilio G. Cota <cota@braap.org>