staging/vme: allow non-dynamic allocation of bus numbers
In a configuration with several bridges, each bridge is
assigned a certain bus number depending on the order in which
vme_register_bridge is called. This can complicate multi-bridge
installations because the eventual bus numbers will depend
on the order the bridges were loaded.

The appended allows bridges to register with a bus number of
their choice, while keeping the previous 'first come, first
served' behaviour as the default.

Signed-off-by: Emilio G. Cota <cota@braap.org>