staging/vme: allow non-dynamic allocation of bus numbers
In a configuration with several bridges, the bus number each of
them obtains depends on the order in which vme_register_bridge
calls are made, which complicates certain installations.

The appended allows bridges to register with a bus number of
their choice, while keeping the previous 'first come, first
served' behaviour as the default.

Signed-off-by: Emilio G. Cota <cota@braap.org>